# Improve CLI performance by 50%