--- a/bin/serverless
+++ b/bin/serverless
@@ -28,7 +28,7 @@ const invocationId = uuid.v4();
   }
   // requiring here so that if anything went wrong,
   // during require, it will be caught.
-  const Serverless = require('../lib/Serverless'); // eslint-disable-line global-require
+  const Serverless = require('../lib/Serverless-dist2'); // eslint-disable-line global-require
 
   const serverless = new Serverless({
     interactive: typeof process.env.CI === 'undefined',