--- a/lib/classes/Config.js
+++ b/lib/classes/Config.js
@@ -2,12 +2,13 @@
 
 const _ = require('lodash');
 const path = require('path');
+const rootPath = require('../../rootPath');
 
 class Config {
 
   constructor(serverless, config) {
     this.serverless = serverless;
-    this.serverlessPath = path.join(__dirname, '..');
+    this.serverlessPath = path.join(rootPath);
 
     if (config) this.update(config);
   }
@@ -18,4 +19,3 @@ class Config {
 }
 
 module.exports = Config;
-