--- a/lib/classes/PluginManager.js
+++ b/lib/classes/PluginManager.js
@@ -8,6 +8,7 @@ const writeFile = require('../utils/fs/writeFile');
 const getCacheFilePath = require('../utils/getCacheFilePath');
 const getServerlessConfigFile = require('../utils/getServerlessConfigFile');
 const crypto = require('crypto');
+const rootPath = require('../../rootPath');
 
 class PluginManager {
   constructor(serverless) {
@@ -92,7 +93,7 @@ class PluginManager {
   }
 
   loadCorePlugins() {
-    const pluginsDirectoryPath = path.join(__dirname, '../plugins');
+    const pluginsDirectoryPath = path.join(rootPath, 'plugins');
     const corePlugins = this.serverless.utils
       .readFileSync(path.join(pluginsDirectoryPath, 'Plugins.json')).plugins
       .map((corePluginPath) => path.join(pluginsDirectoryPath, corePluginPath));