--- a/lib/plugins/aws/invokeLocal/index.js
+++ b/lib/plugins/aws/invokeLocal/index.js
@@ -7,6 +7,7 @@ const validate = require('../lib/validate');
 const chalk = require('chalk');
 const stdin = require('get-stdin');
 const spawn = require('child_process').spawn;
+const rootPath = require('../../../../rootPath');
 
 
 class AwsInvokeLocal {
@@ -152,7 +153,7 @@ class AwsInvokeLocal {
 
     return new BbPromise(resolve => {
       const python = spawn(runtime,
-        [path.join(__dirname, 'invoke.py'), handlerPath, handlerName], { env: process.env });
+        [path.join(rootPath, 'plugins', 'invokeLocal', 'invoke.py'), handlerPath, handlerName], { env: process.env });
       python.stdout.on('data', (buf) => this.serverless.cli.consoleLog(buf.toString()));
       python.stderr.on('data', (buf) => this.serverless.cli.consoleLog(buf.toString()));
       python.stdin.write(input);