--- a/lib/utils/sentry.js
+++ b/lib/utils/sentry.js
@@ -6,12 +6,13 @@ const readFileIfExists = require('./fs/readFileIfExists');
 const getTrackingConfigFileName = require('./getTrackingConfigFileName');
 const path = require('path');
 const BbPromise = require('bluebird');
+const rootPath = require('../../rootPath');
 
 const SLS_DISABLE_ERROR_TRACKING = true;
 const IS_CI = ci.isCI;
 
 function initializeErrorReporter(invocationId) {
-  const trackingConfigFilePath = path.join(__dirname, '..', '..', getTrackingConfigFileName());
+  const trackingConfigFilePath = path.join(rootPath, '..', getTrackingConfigFileName());
   return readFileIfExists(trackingConfigFilePath).then(trackingConfig => {
     const config = configUtils.getConfig();
     const trackingDisabled = config.trackingDisabled;