--- a/rootPath.js
+++ b/rootPath.js
@@ -1,0 +1,1 @@
+module.exports = __dirname;