… lib/Serverless-dist.js
… lib/Serverless-dist2.js
… lib/Serverless-dist3.js