### Goal

Reduce the initial loading times by 50%.

### General Explanation

I started experimenting with a build script to bundle the Serverless framework in less files. The assumption and goal was that I could reduce the initial loading times by 50%.

As a first experiment I managed to get the total time for running the command `sls —help` on my machine from **1.5 seconds to 1 second** using Webpack.

This experiment is in an early stage as our plugin system is written in a way (dynamic imports) that neither rollup nor webpack can handle it as of today. This is fixable by either contributing to the webpack core or what sounds easier to patch the rollup-commonjs-plugin when operating on the AST.

### Idea

The idea is to keep the whole repository as it is. Using `sldev` will still require the whole thing. Only for a production build we create an optimized bundle. The Integration test suite then one more time runs with this file. We patch the bin/serverless file to require the bundle before we publish.

### Why is it faster?

Parsing JavaScript and loading it into the V8 takes time. Having a lot of files can add up to that. Facebook for example working on the project Prepack https://prepack.io/ to speed up thing even if the bundle to download gets a bit larger. Prepack is definitely not production ready, but might be valuable in the future.

Initially I though it would be ideal to have 1 big file with all the possible dependencies, but my initial experiments this isn’t the case. Maybe because stuff is packaged in that’s never used. I simply don’t know yet.

### Next Steps

- [ ] Make dynamic imports working by patching rollup-commonjs or patch Webpack core
- [ ] Update imports to optimise for tree shaking (e.g. const map = require(‘lodash/map’)
- [ ] Investigate packages that can be replaced by slimmer ones e.g. replace apollo with apollo-fetch
- [ ] Minify the output (this should give a solid loading boost!)
- [ ] Investigate source maps for a Node CLI for proper error handling

Based on gut feeling we could even get to 0.7 seconds.

### History

First I started out using rollup as it is known for getting to really good results. Quickly I ran into this error.

```
Dynamic requires are not currently supported by rollup-plugin-commonjs
```

Next up I switched to Webpack. It also missed support for dynamic plugin loading, but I just wanted to figure out if my thesis is valid, excluded the node_modules, made a built, patched the failing dynamic requires in the Webpack bundle output and voila: `sls —help` down from 1.5 seconds down to 1 second.

Further I added the node_modules. The size of the Serverless-dist.js file grew from 160KB (only included all the utils we have) to 2.6MB with all the treeshaked node_modules (not sure how well treeshaking really works with require).

The resulting dist needed 1.5 to 1.6 seconds to run `sls —help`. This was unexpected to me.

In both cases (rollbar & webpack) I had to manually install pg. It was related to ravenjs, but can’t say for sue.

## Measuring

I ran `time sldev —help` where sldev is an alias to the repositories bin. In the bin file I redirected to the build files.

## How did you implement it:

TODO

## How can we verify it:

TODO

## Todos:

- [ ] Write tests
- [ ] Write documentation
- [ ] Fix linting errors
- [ ] Make sure code coverage hasn't dropped
- [ ] Provide verification config / commands / resources
- [ ] Enable "Allow edits from maintainers" for this PR
- [ ] Update the messages below

***Is this ready for review?:*** NO
***Is it a breaking change?:*** NO