@nikgraf 2017-10-10
Closing this since I won't continue working on it.