@pmuens 2017-10-11
> Closing this since I won't continue working on it.

@nikgraf Thanks for taking the time and effort to investigate potential performance enhancements 🙏!